Patches, new yassl, some small fixes
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@623 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ client/HashManager.cpp
@@ -409,7 +409,7 @@ void HashLoader::endTag(const string& name, const string&) {
 
 HashManager::HashStore::HashStore() : dirty(false) 
 { 
-	if(File::getSize(getDataFile()) <= sizeof(int64_t)) {
+	if(File::getSize(getDataFile()) <= static_cast<int64_t>(sizeof(int64_t))) {
 		try {
 			createDataFile(getDataFile());
 		} catch(const FileException&) {
